fix(cli): keep run/ps/logs/push/images/ls library API backward-compatible
The CLI flag-parity work added public fields to the externally-constructible
LsOptions/PsOptions/LogsOptions/PushOptions/RunOverrides structs and changed
images_with_options' arity, which cargo-semver-checks flags as breaking for the
public library crate. Carry the new flag data off the frozen structs instead:
new PsFilterOptions/LogsDisplay options + ps_filtered/logs_with_display/
push_with_quiet/images_with_services/list_projects_filtered, and a run_labels
engine field mirroring run_env_files. Original signatures stay as thin wrappers.

Also fix the Commit --pause merge resolution (duplicate field) and split the
config rendering out of startup into startup/config_render.rs to stay under the
500-line limit.
